Specify Ruby version in .ruby-version file
When developing with Ruby I use `chruby` to switch between different
versions of Ruby. It will automatically switch to the version specified
in `.ruby-version`.

By sharing this config with the production environment we can ensure
consistency. Heroku uses the version specified in `Gemfile` when booting
up the application, and as `Gemfile` is Ruby source, we can just read
the contents of the `.ruby-version` file.