Document the use of `bundle exec` with mr-sparkle
I suspect that back when the docs first mentioned mr-sparkle, it wasn't
part of the bundle. It therefore didn't make sense to tell people to run
it with `bundle exec`.

These days `mr-sparkle` is in Bundler's development group, so we can
expect it to be found when run via bundler. It's far better to use
bundler to run it so that people don't run into problems at startup,
because different versions of other libraries (e.g. rack) get loaded
first.