Bump version of Nesta
Nesta 0.17.0 pulls in sass-embedded 1.80.6, which includes deprecation
of @import and global colour functions. See the Nesta changelog for
for details, but they've necessitated the changes to the Sass here in
order to silence warnings in the server output.

The layout.sass file has been included in master.sass because @use
(which replaces @import) isn't allowed after rules have been declared. I
could have moved the rules in master.sass into another file, but notice
that some layout related rules already exist in master.sass. So why
bother to keep them separate?

If I work on the design of this site again I'll be burning the current
design to the ground, so this feels like a pragmatic approach.